refactor: move `InvocationIdGenerator`
We will need this class in generated libraries that automatically
populate request id fields.
