+++ google/cloud/BUILD.bazel
@@ -102,6 +102,8 @@ to your build command, or set this value in your `.bazelrc` file(s).
         "@boringssl//:crypto",
         "@com_google_absl//absl/base",
         "@com_google_absl//absl/functional:function_ref",
+        "@com_google_absl//absl/strings",
+        "@com_google_absl//absl/strings:str_format",
         "@com_google_absl//absl/time",
         "@com_google_absl//absl/types:optional",
         "@com_google_absl//absl/types:span",
