+++ google/cloud/google_cloud_cpp_common.bzl
@@ -58,6 +58,7 @@ google_cloud_cpp_common_hdrs = [
     "internal/future_then_impl.h",
     "internal/getenv.h",
     "internal/group_options.h",
+    "internal/invocation_id_generator.h",
     "internal/invoke_result.h",
     "internal/ios_flags_saver.h",
     "internal/log_impl.h",
@@ -129,6 +130,7 @@ google_cloud_cpp_common_srcs = [
     "internal/future_impl.cc",
     "internal/future_then_impl.cc",
     "internal/getenv.cc",
+    "internal/invocation_id_generator.cc",
     "internal/log_impl.cc",
     "internal/make_status.cc",
     "internal/noexcept_action.cc",
