+++ google/cloud/google_cloud_cpp_common.cmake
@@ -87,6 +87,8 @@ add_library(
     internal/getenv.cc
     internal/getenv.h
     internal/group_options.h
+    internal/invocation_id_generator.cc
+    internal/invocation_id_generator.h
     internal/invoke_result.h
     internal/ios_flags_saver.h
     internal/log_impl.cc
@@ -174,6 +176,7 @@ target_link_libraries(
            absl::memory
            absl::optional
            absl::span
+           absl::str_format
            absl::time
            absl::variant
            Threads::Threads
@@ -352,6 +355,7 @@ if (BUILD_TESTING)
         internal/future_impl_test.cc
         internal/future_then_impl_test.cc
         internal/group_options_test.cc
+        internal/invocation_id_generator_test.cc
         internal/invoke_result_test.cc
         internal/log_impl_test.cc
         internal/make_status_test.cc
