+++ google/cloud/google_cloud_cpp_common_unit_tests.bzl
@@ -42,6 +42,7 @@ google_cloud_cpp_common_unit_tests = [
     "internal/future_impl_test.cc",
     "internal/future_then_impl_test.cc",
     "internal/group_options_test.cc",
+    "internal/invocation_id_generator_test.cc",
     "internal/invoke_result_test.cc",
     "internal/log_impl_test.cc",
     "internal/make_status_test.cc",
