+++ google/cloud/storage/google_cloud_cpp_storage.bzl
@@ -77,7 +77,6 @@ google_cloud_cpp_storage_hdrs = [
     "internal/hmac_key_requests.h",
     "internal/http_response.h",
     "internal/impersonate_service_account_credentials.h",
-    "internal/invocation_id_generator.h",
     "internal/lifecycle_rule_parser.h",
     "internal/logging_stub.h",
     "internal/make_jwt_assertion.h",
@@ -191,7 +190,6 @@ google_cloud_cpp_storage_srcs = [
     "internal/hmac_key_requests.cc",
     "internal/http_response.cc",
     "internal/impersonate_service_account_credentials.cc",
-    "internal/invocation_id_generator.cc",
     "internal/lifecycle_rule_parser.cc",
     "internal/logging_stub.cc",
     "internal/make_jwt_assertion.cc",
