+++ google/cloud/storage/google_cloud_cpp_storage.cmake
@@ -121,8 +121,6 @@ add_library(
     internal/http_response.h
     internal/impersonate_service_account_credentials.cc
     internal/impersonate_service_account_credentials.h
-    internal/invocation_id_generator.cc
-    internal/invocation_id_generator.h
     internal/lifecycle_rule_parser.cc
     internal/lifecycle_rule_parser.h
     internal/logging_stub.cc
@@ -249,7 +247,6 @@ target_link_libraries(
     PUBLIC absl::cord
            absl::memory
            absl::strings
-           absl::str_format
            absl::time
            absl::variant
            google-cloud-cpp::common
@@ -472,7 +469,6 @@ if (BUILD_TESTING)
         internal/hmac_key_requests_test.cc
         internal/http_response_test.cc
         internal/impersonate_service_account_credentials_test.cc
-        internal/invocation_id_generator_test.cc
         internal/logging_stub_test.cc
         internal/make_jwt_assertion_test.cc
         internal/metadata_parser_test.cc
