+++ google/cloud/storage/storage_client_unit_tests.bzl
@@ -68,7 +68,6 @@ storage_client_unit_tests = [
     "internal/hmac_key_requests_test.cc",
     "internal/http_response_test.cc",
     "internal/impersonate_service_account_credentials_test.cc",
-    "internal/invocation_id_generator_test.cc",
     "internal/logging_stub_test.cc",
     "internal/make_jwt_assertion_test.cc",
     "internal/metadata_parser_test.cc",
